# reduce memory allocations when using New()