This resolves #13. You may want to review the commits individually.

Note that merging this requires a major version bump (to 2.0.0).