@cespare 2019-01-01
@ongardie-ebay hey, I finally got to this. Sorry for the long delay.

I made some changes on the digest branch and merged that. I changed the base of this PR to a temporary branch just to keep it in the same state for now.

To reduce back-and-forth I made followup changes myself on a different branch. I made a `v2` branch from master and rebased your changes on that. Then I opened #16 against the `v2` branch. Please review that PR.

The changes here will be merged as part of the `v2` branch merge so we can close this PR unless there's anything that's easier to discuss here rather than on #16.